pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Pau Duran <pduran.stratify@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"
psycopg2-binary = "^2.9.10"
openai = "^1.65.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
freezegun = "^1.5.1"


[tool.poetry.group.dev.dependencies]
tox = "^4.24.1"
flake8 = "^7.1.1"
black = "^25.1.0"
pytest = "^8.3.4"
mypy = "^1.15.0"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"
flake8-quotes = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
